# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/businesses": {
      "get": {
        "operationId": "BusinessesController_listPublic",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Business name or description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "verifiedOnly",
            "required": false,
            "in": "query",
            "description": "Only administrator-verified businesses",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "verificationStatus",
            "required": false,
            "in": "query",
            "description": "Filter by verification workflow status",
            "schema": {
              "type": "string",
              "enum": ["UNVERIFIED", "PENDING", "VERIFIED", "REJECTED"]
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["recommended", "popular", "newest"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessSummaryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Browse active local businesses",
        "tags": ["businesses"]
      },
      "post": {
        "description": "Free and immediate — no approval step. Verification is a separate signal an administrator grants later.",
        "operationId": "BusinessesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a business",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/mine": {
      "get": {
        "operationId": "BusinessesController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessSummaryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Businesses this user owns or works for",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/{slug}": {
      "get": {
        "operationId": "BusinessesController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessDetailDto"
                }
              }
            }
          }
        },
        "summary": "Public business profile",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/{id}": {
      "patch": {
        "description": "Changing the name, phone or address of a verified business returns it to pending.",
        "operationId": "BusinessesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a business",
        "tags": ["businesses"]
      },
      "delete": {
        "description": "Soft delete. Its listings are archived so none is left pointing at nothing.",
        "operationId": "BusinessesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a business",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/{id}/staff": {
      "get": {
        "operationId": "BusinessesController_listStaff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessStaffDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Staff on a business",
        "tags": ["businesses"]
      },
      "post": {
        "description": "They must already have a LocZ account — this never sends an SMS invitation.",
        "operationId": "BusinessesController_addStaff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessStaffDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a staff member by phone number",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/{id}/staff/{staffId}": {
      "delete": {
        "operationId": "BusinessesController_removeStaff",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a staff member",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/{id}/verification-request": {
      "post": {
        "description": "Requires a description, address, phone number and opening hours. Repeated pending requests are rejected.",
        "operationId": "BusinessesController_requestVerification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request administrator verification for an owned business",
        "tags": ["businesses"]
      }
    },
    "/api/businesses/{id}/verification": {
      "post": {
        "operationId": "BusinessesController_setVerification",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set verification status (admin)",
        "tags": ["businesses"]
      }
    },
...
  },
  "info": {
    "title": "LocZ API",
    "description": "Location-first local discovery platform. Successful responses are wrapped in `{ success, data }`; errors use `{ success: false, error: { code, message } }`.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    },
    {
      "url": "https://locz.in",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "BusinessHourDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "number",
            "minimum": 0,
            "maximum": 6,
            "description": "0 = Sunday"
          },
          "opensAt": {
            "type": "string",
            "example": "09:30"
          },
          "closesAt": {
            "type": "string",
            "example": "21:00"
          },
          "isClosed": {
            "type": "boolean",
            "default": false,
            "description": "Closed all day"
          }
        },
        "required": ["dayOfWeek", "opensAt", "closesAt"]
      },
      "BusinessSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "cityName": {
            "type": "string"
          },
          "logoUrl": {
            "type": "object"
          },
          "verificationStatus": {
            "type": "string",
            "enum": ["UNVERIFIED", "PENDING", "VERIFIED", "REJECTED"]
          },
          "listingCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "description": {
            "type": "object"
          },
          "addressLine": {
            "type": "object"
          },
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessHourDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "categoryName",
          "cityName",
          "verificationStatus",
          "listingCount",
          "viewCount",
          "hours"
        ]
      },
      "CreateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sri Lakshmi Electronics"
          },
          "categoryId": {
            "type": "string"
          },
          "cityId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "addressLine": {
            "type": "string",
            "example": "Shop 12, Road No 36, Jubilee Hills"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "primaryPhone": {
            "type": "string",
            "example": "+919876543210"
          },
          "whatsappNumber": {
            "type": "string",
            "example": "+919876543210"
          },
          "email": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "example": "https://example.com"
          },
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessHourDto"
            }
          }
        },
        "required": ["name", "categoryId", "cityId"]
      },
      "BusinessDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "cityName": {
            "type": "string"
          },
          "logoUrl": {
            "type": "object"
          },
          "verificationStatus": {
            "type": "string",
            "enum": ["UNVERIFIED", "PENDING", "VERIFIED", "REJECTED"]
          },
          "listingCount": {
            "type": "number"
          },
          "viewCount": {
            "type": "number"
          },
          "description": {
            "type": "object"
          },
          "addressLine": {
            "type": "object"
          },
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessHourDto"
            }
          },
          "latitude": {
            "type": "object"
          },
          "longitude": {
            "type": "object"
          },
          "primaryPhone": {
            "type": "object"
          },
          "whatsappNumber": {
            "type": "object"
          },
          "email": {
            "type": "object"
          },
          "website": {
            "type": "object"
          },
          "isOwner": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "categoryName",
          "cityName",
          "verificationStatus",
          "listingCount",
          "viewCount",
          "hours",
          "isOwner",
          "createdAt"
        ]
      },
      "UpdateBusinessDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Sri Lakshmi Electronics"
          },
          "categoryId": {
            "type": "string"
          },
          "cityId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "addressLine": {
            "type": "string",
            "example": "Shop 12, Road No 36, Jubilee Hills"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "primaryPhone": {
            "type": "string",
            "example": "+919876543210"
          },
          "whatsappNumber": {
            "type": "string",
            "example": "+919876543210"
          },
          "email": {
            "type": "string"
          },
          "website": {
            "type": "string",
            "example": "https://example.com"
          },
          "hours": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BusinessHourDto"
            }
          }
        }
      },
      "BusinessStaffDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptedAt": {
            "type": "object"
          }
        },
        "required": ["id", "userId", "displayName", "role", "permissions"]
      },
      "AddStaffDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+919876543210",
            "description": "The staff member must already have a LocZ account"
          },
          "role": {
            "type": "string",
            "enum": ["MANAGER", "EDITOR", "VIEWER"]
          }
        },
        "required": ["phone", "role"]
      },
...
    }
  }
}